refactor: Swagger 어노테이션과 이름 겹칩으로 인해 공통응답구조 파일이름 변경
